backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
openpyxl>=3.1.0

# Statistical analysis
statsmodels>=0.14.0
lifelines>=0.27.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0

# Utilities
httpx>=0.25.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
